Introduce `MeldedMethod`
A `MeldedMethod` tracks the provenance of its components. Without it, we
are unable to determine the difference between:
  - a method that contains a request and its corresponding 4xx response,
    and
  - the result of melding (1) a method that contains a request with no
    response and (2) a method that contains a request and its
    corresponding 4xx response.

In the second scenario, the melded method contains the bare request
(because it might contain useful information) and the 4xx response (but
not its request, which is likely bogus). We wish to retain all aspects
of this melded method when doing further melds, whereas when melding the
method in the first scenario, we might drop its request. However, these
two are indistinguishable, unless we are able to track the provenance of
the components in a melded method.

See the "no response, 4xx example, full non-4xx" test case in
`spec_util/meld_test.go`.