When melding primitives, join up the type lattice
Right now, melding an int32 primitive with another int32 primitive with a data
format will result in a conflict.  However, this should simply result in int32.

The same is true for other base types, and also between different kinds of base
types.  As another example, melding int32 and uint32 should result in int64,
not a conflict.