Extend map heuristics (#138)
We use heuristics to decide when a JSON object in an API is a map
(dictionary) with variable data keys rather than a fixed part of the
API.

This PR extends the heuristics to account for objects with fields that
begin with numbers -- this covers common data formats like numbers,
datetimes, etc.  If every field in a struct beings with a number and
there are sufficient fields, we consider it to be a map.