Distinguish nullable from optional fields (#171)
The Akita IR has an Optional node that is one of Data (the field with this
type is optional, i.e. may be present or absent), or None (the field with this
type has only been seen with "null" as a value).

The Data node also has a "nullable" property, which indicates that the field
with this type is capable of carrying "null" as a value.

Previously, we were treating the None node as optional, rather than nullable.
That is, `meld(None, T) = Optional<T>`.  This PR fixes that, changing melding
so that `meld(None, T) = Nullable<T>` where `Nullable<T>` means that `T.nullable = true`.

It also propagates `Data.nullable` during melding.