Fix bug introduced in #185
This addresses the following issue in `meldOneOf`. Suppose the source and
destination both have multiple variants. During melding, a source variant
can collapse the destination into a singleton. Previously,
`meldOneOfVariant` would eagerly convert the destination into a
Primitive. But if there are more source variants to meld, the next call
to `meldOneOfVariant` would then yield an error because the destination
is no longer a OneOf.

To address this, `meldOneOfVariant` no longer eagerly unwraps singleton
OneOfs. Instead, this is now the caller's responsibility. The function
`rewriteSingletonOneOf` is provided for this purpose.