[POA-1519] Allow HTTP response without reason phrase (#214)
Technically, the RFC requires a space even if the reason phrase is
empty, but at least one service has been observed to violate this. Add
unit tests for both correct and incorrect blank reason phrase, and a one
reproducing the original error to verify that Go's http parser is fine
with this.