Allow PacketCounts to be reset (#230)
I ran into a couple of roadblocks with previous plan for calculating per
agent throughput and total throughput with just the cumulative count. I
have settled on the following plan:

- Introduce a new set of counters, `ObservationWindow` though open to
alternative names, that reset each time we send a `PacketCountSummary`.
- Use the counts in the observation window to build a time series table.
- Query the time series table over a time period to determine the number
of agents that reported rate limited request and sum the number of
requests processed in that time period.